+++ lib/chef/knife/vagrant_server_create.rb
@@ -157,6 +157,10 @@ class Chef
         :long => "--vb-customize VBOXMANAGE_COMMANDS",
         :description => "List of customize options for the virtualbox vagrant provider separated by ::"
 
+      option :vmx_customize,
+        :long => "--vmx-customize VMX_PARAMETERS",
+        :description => "List of parameters for the vmware vagrant providers separated by ::"
+
       option :provider,
         :long => "--provider Provider",
         :description => "The vagrant provider to use for the server",
@@ -208,6 +212,10 @@ class Chef
         customize.split(/::/).collect { |k| "vb.customize [ #{k} ]" }.join("\n") if customize
       end
 
+      def build_vmx_customize(customize)
+        customize.split(/::/).collect { |k| "v.vmx [ #{k} ]" }.join("\n") if customize
+      end
+
       def build_shares(share_folders)
         share_folders.collect do |share|
           host, guest = share.chomp.split "::"
@@ -238,11 +246,13 @@ Vagrant.configure("2") do |config|
   end
   
   config.vm.provider :vmware_fusion do |v|
-     v.vmx["memsize"] = "#{@server.memsize}"
+    v.vmx["memsize"] = "#{@server.memsize}"
+    #{build_vmx_customize(@server.vmx_customize)}
   end
 
   config.vm.provider :vmware_workstation do |v|
-     v.vmx["memsize"] = "#{@server.memsize}"
+    v.vmx["memsize"] = "#{@server.memsize}"
+    #{build_vmx_customize(@server.vmx_customize)}
   end
 
   #{additions.join("\n")}
@@ -322,7 +332,8 @@ end
           :share_folders => config[:share_folders],
           :port_forward => config[:port_forward],
           :use_cachier => config[:use_cachier],
-          :vb_customize => locate_config_value(:vb_customize)
+          :vb_customize => locate_config_value(:vb_customize),
+          :vmx_customize => locate_config_value(:vmx_customize)
         }
 
         # Get specified IP address for new instance or pick an unused one from the subnet pool.